# js/apps/devtools-app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KotoneBot DevTools",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/fs/rename_document/precheck": {
      "post": {
        "tags": [
          "document"
        ],
        "summary": "Precheck Rename Document",
        "operationId": "precheck_rename_document_api_fs_rename_document_precheck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecheckRenameDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RenameDocumentPrecheckResultModel_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/rename_document/execute": {
      "post": {
        "tags": [
          "document"
        ],
        "summary": "Execute Rename Document",
        "operationId": "execute_rename_document_api_fs_rename_document_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRenameDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RenameDocumentExecuteResultModel_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/document/create": {
      "post": {
        "tags": [
          "document"
        ],
        "summary": "Create Document",
        "operationId": "create_document_api_document_create_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_api_document_create_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_CreateDocumentResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/suggest_path": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Suggest Document Path",
        "operationId": "suggest_document_path_api_ai_suggest_path_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_suggest_document_path_api_ai_suggest_path_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_SuggestPathResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/infer_definitions": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Infer Definitions",
        "operationId": "infer_definitions_api_ai_infer_definitions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_infer_definitions_api_ai_infer_definitions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_InferDefinitionsResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/device/adb/screenshot": {
      "get": {
        "tags": [
          "device"
        ],
        "summary": "Capture Adb Screenshot",
        "operationId": "capture_adb_screenshot_api_device_adb_screenshot_get",
        "parameters": [
          {
            "name": "serial",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Serial"
            }
          },
          {
            "name": "displayId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Displayid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_CaptureScreenshotResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversion/execute": {
      "post": {
        "tags": [
          "conversion"
        ],
        "summary": "Conversion Execute",
        "operationId": "conversion_execute_api_conversion_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteConversionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ConversionExecuteResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversion/scan": {
      "post": {
        "tags": [
          "conversion"
        ],
        "summary": "Conversion Scan",
        "operationId": "conversion_scan_api_conversion_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ScanStartResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversion/scan_progress/{task_id}": {
      "get": {
        "tags": [
          "conversion"
        ],
        "summary": "Conversion Scan Progress",
        "operationId": "conversion_scan_progress_api_conversion_scan_progress__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ScanProgress_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversion/scan/{task_id}": {
      "delete": {
        "tags": [
          "conversion"
        ],
        "summary": "Conversion Cancel Scan",
        "operationId": "conversion_cancel_scan_api_conversion_scan__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/server/execute_command": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Execute Server Command",
        "operationId": "execute_server_command_api_server_execute_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_Union_MetaRefetchResult__SymbolUpdateResultModel__RenameDocumentPrecheckResultModel__RenameDocumentExecuteResultModel__RenameSymbolPrecheckResult__RenameSymbolExecuteResult__VariantCloneToImageResult__VariantImportImageResult__VariantCopySelectedPrefabPrecheckResult__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_document_api_document_create_post": {
        "properties": {
          "targetPath": {
            "type": "string",
            "title": "Targetpath"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "targetPath",
          "image"
        ],
        "title": "Body_create_document_api_document_create_post"
      },
...
      "Body_infer_definitions_api_ai_infer_definitions_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "definitionsJson": {
            "type": "string",
            "title": "Definitionsjson"
          },
          "imagePath": {
            "type": "string",
            "title": "Imagepath"
          },
          "providerType": {
            "type": "string",
            "title": "Providertype"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "apiKey": {
            "type": "string",
            "title": "Apikey",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "image",
          "definitionsJson",
          "imagePath",
          "providerType"
        ],
        "title": "Body_infer_definitions_api_ai_infer_definitions_post"
      },
...
      "Body_suggest_document_path_api_ai_suggest_path_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "providerType": {
            "type": "string",
            "title": "Providertype"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "apiKey": {
            "type": "string",
            "title": "Apikey",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "image",
          "providerType"
        ],
        "title": "Body_suggest_document_path_api_ai_suggest_path_post"
      },
...
      "ExecuteConversionRequest": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/ConfirmedMatch"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "matches"
        ],
        "title": "ExecuteConversionRequest"
      },
      "ExecuteRenameDocumentRequest": {
        "properties": {
          "sourceImagePath": {
            "type": "string",
            "title": "Sourceimagepath"
          },
          "targetImagePath": {
            "type": "string",
            "title": "Targetimagepath"
          }
        },
        "type": "object",
        "required": [
          "sourceImagePath",
          "targetImagePath"
        ],
        "title": "ExecuteRenameDocumentRequest",
        "description": "文档重命名执行请求。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrecheckRenameDocumentRequest": {
        "properties": {
          "sourceImagePath": {
            "type": "string",
            "title": "Sourceimagepath"
          },
          "targetImagePath": {
            "type": "string",
            "title": "Targetimagepath"
          }
        },
        "type": "object",
        "required": [
          "sourceImagePath",
          "targetImagePath"
        ],
        "title": "PrecheckRenameDocumentRequest",
        "description": "文档重命名预检请求。"
      },
...
      "RenameDocumentPrecheckResultModel": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/RenameDocumentItemModel"
            },
            "type": "array",
            "title": "Documents"
          },
          "fileRenames": {
            "items": {
              "$ref": "#/components/schemas/RenameFileItemModel"
            },
            "type": "array",
            "title": "Filerenames"
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "hasConflicts": {
            "type": "boolean",
            "title": "Hasconflicts"
          }
        },
        "type": "object",
        "required": [
          "documents",
          "fileRenames",
          "conflicts",
          "hasConflicts"
        ],
        "title": "RenameDocumentPrecheckResultModel",
        "description": "文档重命名预检结果。"
      },
...
      "ResponseModel_CaptureScreenshotResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptureScreenshotResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[CaptureScreenshotResponse]"
      },
      "ResponseModel_ConversionExecuteResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConversionExecuteResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[ConversionExecuteResponse]"
      },
...
      "ResponseModel_CreateDocumentResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreateDocumentResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[CreateDocumentResult]"
      },
...
      "ResponseModel_InferDefinitionsResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferDefinitionsResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[InferDefinitionsResult]"
      },
...
      "ResponseModel_NoneType_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[NoneType]"
      },
...
      "ResponseModel_RenameDocumentExecuteResultModel_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenameDocumentExecuteResultModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[RenameDocumentExecuteResultModel]"
      },
      "ResponseModel_RenameDocumentPrecheckResultModel_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenameDocumentPrecheckResultModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[RenameDocumentPrecheckResultModel]"
      },
...
      "ResponseModel_ScanProgress_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[ScanProgress]"
      },
      "ResponseModel_ScanStartResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanStartResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[ScanStartResponse]"
      },
      "ResponseModel_SuggestPathResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SuggestPathResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[SuggestPathResponse]"
      },
...
      "ScanRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "imagePaths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagepaths"
          },
          "screenshotPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screenshotpath"
          },
          "singleImagePath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Singleimagepath"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ScanRequest",
        "description": "启动扫描请求。"
      },
...
    }
  }
}